# apps/api/src/shared-equipment/dto/update-group-activity.dto.ts
import { PartialType } from '@nestjs/swagger';
import { CreateGroupActivityDto } from './create-group-activity.dto';

export class UpdateGroupActivityDto extends PartialType(CreateGroupActivityDto) {}