# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/groups/{groupId}": {
      "get": {
        "operationId": "GroupsController_getGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get group details (members only)",
        "tags": [
          "Groups"
        ]
      },
      "patch": {
        "operationId": "GroupsController_updateGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update group (OWNER/ADMIN only)",
        "tags": [
          "Groups"
        ]
      },
      "delete": {
        "operationId": "GroupsController_deleteGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete group (OWNER only)",
        "tags": [
          "Groups"
        ]
      }
    },
...
    "/groups/{groupId}/activities": {
      "post": {
        "operationId": "GroupActivitiesController_createActivity",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupActivityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a group activity",
        "tags": [
          "Group Activities"
        ]
      },
      "get": {
        "operationId": "GroupActivitiesController_listActivities",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "PUBLISHED",
                "CANCELLED",
                "COMPLETED",
                "ARCHIVED"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive title or venue search",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List group activities",
        "tags": [
          "Group Activities"
        ]
      }
    },
    "/groups/{groupId}/activities/{activityId}": {
      "get": {
        "operationId": "GroupActivitiesController_getActivity",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get activity with shared items",
        "tags": [
          "Group Activities"
        ]
      },
      "patch": {
        "operationId": "GroupActivitiesController_updateActivity",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupActivityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an event and its venue or lifecycle details",
        "tags": [
          "Group Activities"
        ]
      }
    },
    "/groups/{groupId}/activities/{activityId}/duplicate": {
      "post": {
        "operationId": "GroupActivitiesController_duplicateActivity",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Duplicate an event as a new draft",
        "tags": [
          "Group Activities"
        ]
      }
    },
    "/groups/{groupId}/activities/{activityId}/archive": {
      "post": {
        "operationId": "GroupActivitiesController_archiveActivity",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Archive an event while retaining its history",
        "tags": [
          "Group Activities"
        ]
      }
    },
...
    "/groups/{groupId}/activities/{activityId}/weather": {
      "get": {
        "operationId": "WeatherController_get",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Weather Preparation"
        ]
      }
    },
    "/groups/{groupId}/activities/{activityId}/weather/refresh": {
      "post": {
        "operationId": "WeatherController_refresh",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Refresh and evaluate the event-time forecast",
        "tags": [
          "Weather Preparation"
        ]
      }
    },
    "/groups/{groupId}/activities/{activityId}/weather-suggestions/{suggestionId}/accept": {
      "post": {
        "operationId": "WeatherController_accept",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "suggestionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewWeatherSuggestionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Weather Preparation"
        ]
      }
    },
    "/groups/{groupId}/activities/{activityId}/weather-suggestions/{suggestionId}/dismiss": {
      "post": {
        "operationId": "WeatherController_dismiss",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "suggestionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Weather Preparation"
        ]
      }
    },
    "/groups/{groupId}/weather-rules/{ruleKey}": {
      "patch": {
        "operationId": "WeatherController_updateRule",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ruleKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWeatherRuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Weather Preparation"
        ]
      }
    },
...
  },
  "info": {
    "title": "PackPlay API",
    "description": "PackPlay backend API for group equipment management",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Group name",
            "example": "Sunday Football"
          },
          "description": {
            "type": "string",
            "description": "Group description",
            "example": "Weekly football meetup"
          },
          "sportType": {
            "type": "string",
            "description": "Sport type",
            "example": "football"
          }
        }
      },
...
      "CreateGroupActivityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200
          },
          "activityType": {
            "type": "string",
            "enum": [
              "TRAINING",
              "COMPETITION",
              "CASUAL",
              "TRAVEL"
            ]
          },
          "sportProfileId": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "endAt": {
            "type": "string",
            "format": "date-time"
          },
          "venueName": {
            "type": "string",
            "maxLength": 200
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "environment": {
            "type": "string",
            "enum": [
              "INDOOR",
              "COVERED_OUTDOOR",
              "OUTDOOR"
            ]
          },
          "surface": {
            "type": "string",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 5000
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "CANCELLED",
              "COMPLETED",
              "ARCHIVED"
            ],
            "default": "DRAFT"
          },
          "responsibilityDeadline": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "name",
          "activityType"
        ]
      },
      "UpdateGroupActivityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200
          },
          "activityType": {
            "type": "string",
            "enum": [
              "TRAINING",
              "COMPETITION",
              "CASUAL",
              "TRAVEL"
            ]
          },
          "sportProfileId": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "endAt": {
            "type": "string",
            "format": "date-time"
          },
          "venueName": {
            "type": "string",
            "maxLength": 200
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "environment": {
            "type": "string",
            "enum": [
              "INDOOR",
              "COVERED_OUTDOOR",
              "OUTDOOR"
            ]
          },
          "surface": {
            "type": "string",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 5000
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "CANCELLED",
              "COMPLETED",
              "ARCHIVED"
            ],
            "default": "DRAFT"
          },
          "responsibilityDeadline": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ReviewWeatherSuggestionDto": {
        "type": "object",
        "properties": {
          "target": {
            "type": "string",
            "enum": [
              "PERSONAL",
              "SHARED"
            ]
          },
          "itemName": {
            "type": "string",
            "maxLength": 200
          },
          "quantity": {
            "type": "number",
            "minimum": 1
          },
          "checklistId": {
            "type": "string",
            "format": "uuid",
            "description": "Required when accepting as a personal item"
          }
        }
      },
      "UpdateWeatherRuleDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ]
      },
...
    }
  }
}